Move entity classes to the `Entities` namespace
